refactor: split widgets/font_loader/svg_loader/html_renderer/changelog/about_info into directory modules (Tasks 6.15-6.22)
